fix: rename `slot_iterator` into `slots_iterator` to avoid naming conflict with Slot Iterator
