+++ src/models/slots_iterator/iterator.rs
