+++ src/models/slots_iterator/utils.rs
