+++ src/tests/general.rs
@@ -8,7 +8,7 @@ use std::vec;
 
 #[test]
 fn get_next_monday() {
-    use crate::models::slot_iterator::utils::get_start_of_repeat_step;
+    use crate::models::slots_iterator::utils::get_start_of_repeat_step;
 
     let monday = NaiveDate::from_ymd_opt(2022, 09, 26)
         .unwrap()
@@ -44,7 +44,7 @@ fn get_next_monday() {
 
 #[test]
 fn get_next_weekend() {
-    use crate::models::slot_iterator::utils::get_start_of_repeat_step;
+    use crate::models::slots_iterator::utils::get_start_of_repeat_step;
 
     let monday = NaiveDate::from_ymd_opt(2022, 09, 26)
         .unwrap()
